Changed the functions to switch and added some buttons
Now the functions are called by a switch function, added some buttons
to set the color to white when the mouse hover, the fade button now
add .1 to the rgba in the a-value, but if the color is different than black
(0,0,0, a) it sets it to (0,0,0,.1). Added new functions to the buttons, all of them
ends in 2 (because they were created in the second attempt). So new functions, new menu
to call the functions, added the paint white, and set the fade to (0,0,0,.1) if the
square wasn't black.

various adds, i had to create a new branch. i'll merge it later.